+++ test/Sentry.Testing/VerifyExtensions.cs
@@ -32,7 +32,7 @@ public static class VerifyExtensions
                 _ => _.Headers)
             .IgnoreMembers<SessionUpdate>(
                 _ => _.Duration)
-            .IgnoreMembers<Transaction>(
+            .IgnoreMembers<SentryTransaction>(
                 _ => _.Release)
             .IgnoreMembers<SentryException>(
                 _ => _.Module,